feat: Add MLRemoteModel class for remote ML models
The code changes introduce a new class, `MLRemoteModel`, which is responsible for handling remote ML models. This class is added to support the functionality of downloading and using ML models offline. The addition of this class is in line with the recent commits that have been made to create download model classes for offline use and local translation.